+++ postproc/maskload.py
@@ -107,7 +107,7 @@ def SUB(sub):
 def read_Positions_for_Pointprofiles(filename):
 
     dtIN  = np.dtype([('Name','U100'), ('Lon',np.float32), ('Lat',np.float32)])
-    dtOUT = np.dtype([('Name','U100'), ('Lon',np.float32), ('Lat',np.float32), ('i',np.int), ('j',np.int)])
+    dtOUT = np.dtype([('Name','U100'), ('Lon',np.float32), ('Lat',np.float32), ('i',int), ('j',int)])
     MeasPoints=np.loadtxt(filename, dtype=dtIN, skiprows=1,ndmin=1,delimiter="\t")
     
     nMeas=MeasPoints.size
